refactor: helper함수를 utils로 이름 변경
