+++ packages/swagger-generator/src/utils/template.ts
